Moved XXX_BREAKPOINT definitions from App.vue to BreakPoints.ts.
