+++ src/BreakPoints.ts
@@ -1,0 +1,6 @@
+export const XLARGE_BREAKPOINT = 1450
+export const LARGE_BREAKPOINT = 1280
+export const MEDIUM_BREAKPOINT = 1080
+export const SMALL_BREAKPOINT = 768
+export const FINAL_BREAKPOINT = 640
+export const ORUGA_MOBILE_BREAKPOINT = "1080px"
