+++ src/components/account/NftsTable.vue
@@ -116,7 +116,6 @@
 
 import {onBeforeUnmount, onMounted, PropType, watch} from 'vue';
 import {Nft, Token} from "@/schemas/HederaSchemas";
-import {ORUGA_MOBILE_BREAKPOINT} from '@/App.vue';
 import EmptyTable from "@/components/EmptyTable.vue";
 import {routeManager} from "@/router";
 import NftCell, {NftCellItem} from "@/components/token/NftCell.vue";
@@ -125,6 +124,7 @@ import TokenCell, {TokenCellItem} from "@/components/token/TokenCell.vue";
 import {AppStorage} from "@/AppStorage";
 import TablePageSize from "@/components/transaction/TablePageSize.vue";
 import TokenIOL from "@/components/values/link/TokenIOL.vue";
+import {ORUGA_MOBILE_BREAKPOINT} from "@/BreakPoints";
 
 const props = defineProps({
   controller: {
