+++ src/components/allowances/NftAllSerialsAllowanceTable.vue
@@ -97,7 +97,7 @@
 
 import {computed, ComputedRef, defineComponent, inject, PropType, ref, Ref, watch} from 'vue';
 import {NftAllowance} from "@/schemas/HederaSchemas";
-import {ORUGA_MOBILE_BREAKPOINT} from '@/App.vue';
+import {ORUGA_MOBILE_BREAKPOINT} from "@/BreakPoints";
 import TimestampValue from "@/components/values/TimestampValue.vue";
 import EmptyTable from "@/components/EmptyTable.vue";
 import AccountLink from "@/components/values/link/AccountLink.vue";
