+++ src/components/allowances/TokenAllowanceTable.vue
@@ -101,7 +101,7 @@
 import {computed, ComputedRef, defineComponent, inject, PropType, ref, Ref, watch} from 'vue';
 import {TokenAllowance} from "@/schemas/HederaSchemas";
 import {isValidAssociation} from "@/schemas/HederaUtils";
-import {ORUGA_MOBILE_BREAKPOINT} from '@/App.vue';
+import {ORUGA_MOBILE_BREAKPOINT} from "@/BreakPoints";
 import TimestampValue from "@/components/values/TimestampValue.vue";
 import EmptyTable from "@/components/EmptyTable.vue";
 import AccountLink from "@/components/values/link/AccountLink.vue";
