+++ src/components/token/TokenBalanceTable.vue
@@ -87,7 +87,7 @@ import {ComputedRef, defineComponent, inject, PropType, Ref} from 'vue';
 import {TokenDistribution} from "@/schemas/HederaSchemas";
 import {routeManager} from "@/router";
 import TokenAmount from "@/components/values/TokenAmount.vue";
-import {ORUGA_MOBILE_BREAKPOINT} from '@/App.vue';
+import {ORUGA_MOBILE_BREAKPOINT} from "@/BreakPoints";
 import EmptyTable from "@/components/EmptyTable.vue";
 import {TokenBalanceTableController} from "@/components/token/TokenBalanceTableController";
 import AccountIOL from "@/components/values/link/AccountIOL.vue";
