+++ src/components/token/TokenTable.vue
@@ -86,7 +86,7 @@
 import {ComputedRef, defineComponent, inject, PropType, Ref} from 'vue';
 import {routeManager} from "@/router";
 import {Token} from "@/schemas/HederaSchemas";
-import {ORUGA_MOBILE_BREAKPOINT} from '@/App.vue';
+import {ORUGA_MOBILE_BREAKPOINT} from "@/BreakPoints";
 import EmptyTable from "@/components/EmptyTable.vue";
 import {TokenTableController} from "@/components/token/TokenTableController";
 import TokenIOL from "@/components/values/link/TokenIOL.vue";
